+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =====
 
 * (bug) Do not modify original UnitOfWork in `DoctrineChangeChecker`.
+* (improvement) Allow to automatically redact changed fields' values in `DoctrineChangeChecker::getEntityChanges()`.
 
 
 2.5.1